[APDV-113] adding CI/CD and removing linter warnings
