+++ .github/workflows/frontend_test.yml
@@ -21,6 +21,7 @@ jobs:
       - run: flutter config --enable-web
       - run: touch .env
       - run: flutter build web
+      - run: flutter analyze
 
   build_apk:
     name: Build Flutter (Android)
@@ -42,3 +43,4 @@ jobs:
       - run: flutter pub get
       - run: touch .env
       - run: flutter build appbundle --no-tree-shake-icons
+      - run: flutter analyze
