+++ frontend/lib/Models/endpoint_view_provider.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
-import '../DataModels/EndpointData.dart';
+import '../DataModels/endpoint_data.dart';
 
 class DataTypeTab {
   DataTypeTab({required this.typeName});
