+++ frontend/lib/Repository/EndpointRepository/endpoint_rest_repository.dart
@@ -1,10 +1,10 @@
-import 'package:adpv_frontend/DataModels/EndpointSummary.dart';
+import 'package:adpv_frontend/DataModels/endpoint_summary.dart';
 import 'package:dio/dio.dart';
 
-import '../../Common/Consts.dart';
-import '../../Common/URLs.dart';
-import '../../DataModels/EnableField.dart';
-import '../../DataModels/EndpointData.dart';
+import '../../Common/consts.dart';
+import '../../Common/urls.dart';
+import '../../DataModels/enable_field.dart';
+import '../../DataModels/endpoint_data.dart';
 
 class BackendResponse<T> {
   final T data;
@@ -33,13 +33,17 @@ class EndpointRestRepository {
         if (backendResponse.error == "") {
           List<EndpointSummary> endpointSummaryList = [];
           endpointSummaryList = backendResponse.data
-              .map<EndpointSummary>((e) =>
-                  EndpointSummary.fromJson(e)) // do not refactor! UFO MAGIC!
+              .map<EndpointSummary>(
+            // ignore: unnecessary_lambdas
+                (e) => EndpointSummary.fromJson(e),
+              ) // do not refactor! UFO MAGIC!
               .toList();
           return Future.value(endpointSummaryList);
         }
       }
-    } catch (error) {}
+    } on Exception catch (error) {
+      return Future.error(error);
+    }
     return Future.value([]);
   }
 
@@ -53,15 +57,19 @@ class EndpointRestRepository {
     limit = limit ?? 25;
     offset = offset ?? 0;
     try {
-      final Response response =
-          await client.get(backendURL + getEndpointDataURL, queryParameters: {
-        'sensorId': id,
-        'limit': limit,
-        'offset': offset,
-      });
+      final Response response = await client.get(
+        backendURL + getEndpointDataURL,
+        queryParameters: {
+          'sensorId': id,
+          'limit': limit,
+          'offset': offset,
+        },
+      );
 
-      final Response fields = await client.get(backendURL + getFieldEnable,
-          queryParameters: {'endpointId': id});
+      final Response fields = await client.get(
+        backendURL + getFieldEnable,
+        queryParameters: {'endpointId': id},
+      );
 
       if (response.statusCode == 200 && fields.statusCode == 200) {
         final BackendResponse backendResponse =
@@ -78,8 +86,10 @@ class EndpointRestRepository {
           final EndpointData endpointData = EndpointData(
             backendResponse.data.map<Map<dynamic, dynamic>>((e) {
               final Map map = Map.from(e);
-              map.removeWhere((key, value) =>
-                  _isChartData(key, enableFields) && key != ignoreField);
+              map.removeWhere(
+                (key, value) =>
+                    _isChartData(key, enableFields) && key != ignoreField,
+              );
               return map;
             }).toList(),
             backendResponse.data.map<Map<dynamic, dynamic>>((e) {
@@ -93,7 +103,7 @@ class EndpointRestRepository {
         }
       }
     } on Exception catch (error) {
-      print(error);
+      return Future.error(error);
     }
     return Future(EndpointData.empty);
   }
