+++ frontend/lib/Views/endpoint_view.dart
@@ -1,16 +1,16 @@
-import 'package:adpv_frontend/Repository/EndpointRepository/EndpointGateway.dart';
+import 'package:adpv_frontend/Repository/EndpointRepository/endpoint_gateway.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import '../DataModels/EndpointData.dart';
-import '../Models/EndpointViewProvider.dart';
-import '../Widgets/CommonWidgets.dart';
-import '../Widgets/EndpointView/EndpointInfoTable.dart';
-import '../Widgets/EndpointView/TitledLineChart.dart';
+import '../DataModels/endpoint_data.dart';
+import '../Models/endpoint_view_provider.dart';
+import '../Widgets/EndpointView/endpoint_info_table.dart';
+import '../Widgets/EndpointView/titled_line_chart.dart';
+import '../Widgets/common_widgets.dart';
 
 class EndpointView extends StatefulWidget {
   const EndpointView(
-      {required this.endpointId, required this.endpointGateway, Key? key})
+      {required this.endpointId, required this.endpointGateway, Key? key,})
       : super(key: key);
   final int endpointId;
   final EndpointGateway endpointGateway;
@@ -20,27 +20,27 @@ class EndpointView extends StatefulWidget {
 }
 
 TabBar _buildTabBar(
-        EndpointViewProvider endpointViewProvider, BuildContext context) =>
+        EndpointViewProvider endpointViewProvider, BuildContext context,) =>
     TabBar(
         isScrollable: true,
         padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 20),
         unselectedLabelStyle: const TextStyle(
-            color: Colors.pink, fontFamily: "SofiaSans", fontSize: 20),
+            color: Colors.pink, fontFamily: "SofiaSans", fontSize: 20,),
         labelColor: Colors.pink,
         unselectedLabelColor: const Color.fromARGB(255, 127, 166, 168),
         labelStyle: const TextStyle(
-            color: Colors.pink, fontFamily: "SofiaSans", fontSize: 25),
+            color: Colors.pink, fontFamily: "SofiaSans", fontSize: 25,),
         tabs: endpointViewProvider.tabs
             .map((data) => SizedBox(
                   width: MediaQuery.of(context).size.width *
                       1 /
                       endpointViewProvider.tabs.length,
                   child: Tab(text: data.typeName),
-                ))
-            .toList());
+                ),)
+            .toList(),);
 
 SizedBox _buildBarView(EndpointViewProvider endpointViewProvider,
-        BuildContext context, snapshot) =>
+        BuildContext context, snapshot,) =>
     SizedBox(
       height: MediaQuery.of(context).size.height * 0.5,
       child: TabBarView(
@@ -49,12 +49,12 @@ SizedBox _buildBarView(EndpointViewProvider endpointViewProvider,
                     chartName: dataKey.typeName +
                         spacer +
                         endpointViewProvider.getChartUnitName(
-                            dataKey.typeName, snapshot.data),
+                            dataKey.typeName, snapshot.data,),
                     measureFnCallback: (Map<dynamic, dynamic> dataMap, _) =>
                         dataMap[dataKey.typeName],
                     data: snapshot.data!,
-                  ))
-              .toList()),
+                  ),)
+              .toList(),),
     );
 
 class _EndpointViewState extends State<EndpointView> {
@@ -89,14 +89,15 @@ class _EndpointViewState extends State<EndpointView> {
                                 child: Column(children: [
                                   _buildTabBar(endpointViewProvider, context),
                                   _buildBarView(
-                                      endpointViewProvider, context, snapshot)
-                                ])),
-                      )),
+                                      endpointViewProvider, context, snapshot,)
+                                ],),),
+                      ),),
                 ],
-              )),
-            ));
-      });
+              ),),
+            ),);
+      },);
 
+  // ignore: always_declare_return_types
   _pullDownRefresh() async {
     endpointData = widget.endpointGateway
         .getEndpointData(widget.endpointId, null, null, true);
