+++ frontend/lib/main.dart
@@ -1,9 +1,9 @@
-import 'package:adpv_frontend/Repository/EndpointRepository/EndpointGateway.dart';
-import 'package:adpv_frontend/Repository/UserRepository/UserGateway.dart';
+import 'package:adpv_frontend/Repository/EndpointRepository/endpoint_gateway.dart';
+import 'package:adpv_frontend/Repository/UserRepository/user_gateway.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 
-import 'App.dart';
+import 'app.dart';
 
 Future main() async {
   await loadEnvFile();
